chore: move lesson editor canvas and toolbar into tiles-editor
