+++ packages/tiles-editor/src/index.ts
@@ -2,6 +2,13 @@ export * from './components/TileRenderer';
 export * from './components/RichTextEditor';
 export * from './components/TileFrame';
 export * from './components/shared';
+export * from './components/canvas/LessonCanvas';
+export * from './components/palette/TilePalette';
+export * from './components/toolbar/TopToolbar';
+export * from './components/toolbar/AlignmentControls';
+export * from './components/toolbar/FontSelector';
+export * from './components/toolbar/FontSizeSelector';
+export * from './components/toolbar/TextColorPicker';
 export * from './tiles';
 export * from './side-editors';
 export * from './hooks/useLessonEditor';
