+++ src/Pages/LessonEditor.tsx
@@ -1,10 +1,14 @@
 import React, { useState, useEffect, useRef } from 'react';
 import { ArrowLeft, Save, RotateCcw, Grid, Edit } from 'lucide-react';
 import { Lesson, Course } from '../types/course.ts';
-import { TileSideEditor, useLessonEditor, useLessonContentManager } from 'tiles-editor';
-import { TilePalette } from '../components/admin/editor side/TilePalette.tsx';
-import { LessonCanvas } from '../components/admin/canvas/LessonCanvas.tsx';
-import { TopToolbar } from '../components/admin/editor top/TopToolbar.tsx';
+import {
+  LessonCanvas,
+  TilePalette,
+  TopToolbar,
+  TileSideEditor,
+  useLessonEditor,
+  useLessonContentManager,
+} from 'tiles-editor';
 import { Editor } from '@tiptap/react';
 import { ToastContainer } from '../components/common/Toast.tsx';
 import { useToast } from '../hooks/useToast.ts';
